Update CI workflow for testing improvements and coverage reporting
This commit updates the CI workflow to include support for the Python 3.13.0-alpha.5 version and allows prerelease versions. It also updates Poetry to version 1.8.4, adds steps to display Python and Poetry versions, and lists detailed environment information. Additionally, the coverage report setup is enhanced to include HTML reports and an additional upload step for Linux and Python 3.12 configuration.